% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_usa.R
\name{hex_usa}
\alias{hex_usa}
\title{Creates hexagonal tile map for the states of USA.}
\usage{
hex_usa(d = 0.95, center = FALSE, style = "NPR", size = 1,
  long_offset = 0, lat_offset = 0, exclude = character())
}
\arguments{
\item{d}{Numeric. Number between 0 and 1. Procentwise Diameter (length from
center to corner) of the tiles.}

\item{center}{Logical. When TRUE returns the center coordinates of the tile
map.}

\item{style}{Character. Selets the layout style of the tile map.}

\item{size}{Numeric. Size of the tiles.}

\item{long_offset}{Numeric. Value to offset the long output.}

\item{lat_offset}{Numeric. Value to offset the lat output.}

\item{exclude}{Character. Vector of state names which should be excluded
from the output. Matched with lowercase full state names.}
}
\value{
The available styles for this functions are "NPR".
}
\description{
Creates hexagonal tile map for the states of USA.
}
\examples{
library(ggplot2)
crimes <- data.frame(state = tolower(rownames(USArrests)), USArrests)
states_map <- hex_usa()

ggplot(crimes, aes(map_id = state)) +
  geom_map(aes(fill = Murder), map = states_map) +
  expand_limits(x = states_map$long, y = states_map$lat)

states_map <- hex_usa(d = 0.5)

ggplot(crimes, aes(map_id = state)) +
  geom_map(aes(fill = Murder), map = states_map) +
  expand_limits(x = states_map$long, y = states_map$lat) +
  geom_text(data = hex_usa(d = 0.5, center = TRUE),
            aes(x = long, y = lat, label = states_abb),
            inherit.aes = FALSE)
}
